Mockuser displaying in component
