+++ src/components/connections/Connections.js
@@ -1,59 +1,30 @@
-import React, {useState} from "react";
+import React, { useState, useEffect } from "react";
 import Paper from '@material-ui/core/Paper';
-import PropTypes from 'prop-types';
-import SwipeableViews from 'react-swipeable-views';
-import { useTheme } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Tabs from '@material-ui/core/Tabs';
-import Tab from '@material-ui/core/Tab';
-import Typography from '@material-ui/core/Typography';
-import Box from '@material-ui/core/Box';
-
-function TabPanel(props) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`full-width-tabpanel-${index}`}
-      aria-labelledby={`full-width-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box p={2}>
-          <Typography>{children}</Typography>
-        </Box>
-      )}
-    </div>
-  );
-}
-
-TabPanel.propTypes = {
-  children: PropTypes.node,
-  index: PropTypes.any.isRequired,
-  value: PropTypes.any.isRequired,
-};
-
-function a11yProps(index) {
-  return {
-    id: `full-width-tab-${index}`,
-    'aria-controls': `full-width-tabpanel-${index}`,
-  };
-}
+import SwipeTabs from './SwipeTabs';
+import { axiosWithAuth } from "../apiStuff/axiosWithAuth";
 
 function Connections(props) {
 
-  const theme = useTheme();
-  const [value, setValue] = useState(0);
+  const [profilesToDisplay, setProfilesToDisplay] = useState([]);
+  const [profiles, setProfiles] = useState([]);
 
-  const handleChange = (event, newValue) => {
-    setValue(newValue);
-  };
+  useEffect(() => {
+    const apiCallUsers = async () => {
+      await axiosWithAuth()
+        .get("/profile")
+        .then((results) => {
+          if (results) {
+            let data = results.data;
+            console.log("API response", results.data);
+            setProfiles(data);
+          }
+        })
+        .catch((err) => console.log(err));
+    };
+    apiCallUsers();
+  }, []);
 
-  const handleChangeIndex = (index) => {
-    setValue(index);
-  };
+
 
   return (
 
@@ -64,32 +35,8 @@ function Connections(props) {
           My Connections (Sum)
         </h1>
         <div>
-      <AppBar position="static" color="default">
-        <Tabs
-          value={value}
-          onChange={handleChange}
-          indicatorColor="secondary"
-          textColor="primary"
-          variant="fullWidth"
-          aria-label="tabs"
-        >
-          <Tab label="Mentors (#)" {...a11yProps(0)} />
-          <Tab label="Mentees (#)" {...a11yProps(1)} />
-        </Tabs>
-      </AppBar>
-      <SwipeableViews
-        axis={theme.direction === 'rtl' ? 'x-reverse' : 'x'}
-        index={value}
-        onChangeIndex={handleChangeIndex}
-      >
-        <TabPanel value={value} index={0} dir={theme.direction}>
-          <Typography className='text'>Mentor cards go here</Typography>
-        </TabPanel>
-        <TabPanel value={value} index={1} dir={theme.direction}>
-        <Typography className='text'>Mentee cards go here</Typography>
-        </TabPanel>
-      </SwipeableViews>
-    </div>
+        <SwipeTabs/>
+        </div>
         </Paper>
         </div>
       </div>
